+++ NVEncCore/rgy_hdr10plus.h
@@ -45,7 +45,7 @@ public:
     virtual ~RGYHDR10Plus();
 
     RGY_ERR init(const tstring& inputJson);
-    const std::vector<uint8_t> getData(int iframe);
+    const std::vector<uint8_t> getData(int64_t iframe);
     const tstring &inputJson() const { return m_inputJson; };
     tstring getError();
 protected:
